Extract f-string tests into a temporary test folder for py36 and py37
This is not the best long-term solution, just something that can work
until py38 is released and support for py35 can be dropped.
