+++ tox.ini
@@ -48,6 +48,7 @@ commands =
     lint: make lint
     lintexamples: make lintexamples
     test: pytest {posargs:tests}
+    py3{6,7}-test: pytest tests_py36_plus
 skip_install =
     lint: true
     lintexamples: true
